services/backend/requirements.txt
asyncpg==0.27.0
fastapi==0.88.0
uvicorn==0.20.0
Pillow==10.0.0
numpy==1.23.3
matplotlib==3.7.2
opencv-python==4.7.0.72
PyYAML==6.0.1
requests==2.28.1
scipy==1.11.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
pandas==2.0.3
seaborn==0.12.2
psutil==5.9.5
thop==0.1.1.post2209072238
tensorflow==2.13.0
ultralytics==8.0.208
onnx==1.14.0
onnx2pytorch==0.4.1
python-multipart==0.0.6
tortoise-orm==0.19.2
aerich==0.7.2
py-cpuinfo==9.0.0